feat: CRUD para parceiros e redes sociais

+++ src/main/java/br/org/larescolaredencao/repository/ParceiroRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import br.org.larescolaredencao.model.Parceiro;
 
 
-public interface ParceiroRepository extends JpaRepository<Parceiro, Integer> {
+public interface ParceiroRepository extends JpaRepository<Parceiro, Long> {
 	
 }
